+++ src/main/java/com/broscraft/cda/observers/OrderObserver.java
@@ -1,5 +1,7 @@
 package com.broscraft.cda.observers;
 
+import java.math.BigDecimal;
+
 import com.broscraft.cda.dtos.orders.OrderDTO;
 import com.broscraft.cda.dtos.orders.input.NewOrderDTO;
 import com.broscraft.cda.dtos.transaction.TransactionSummaryDTO;
@@ -7,5 +9,5 @@ import com.broscraft.cda.dtos.transaction.TransactionSummaryDTO;
 public interface OrderObserver {
     public void onNewOrder(NewOrderDTO newOrderDTO);
     public void onFillOrder(TransactionSummaryDTO transactionSummaryDTO);
-    public void onRemoveOrder(OrderDTO orderDTO, Float nextBestPrice);
+    public void onRemoveOrder(OrderDTO orderDTO, BigDecimal nextBestPrice);
 }
